Move SignInUp modal to MainHeader component and refactor it
