+++ src/Training.AirBnb.Clone.Frontend/src/App.vue
@@ -1,32 +1,12 @@
 <template>
 
-  <!--    <main-logo/>-->
-  <home-view @sign-in-up-request="toggleModal"/>
-
-  <sign-in-up-modal :modal-active="modalActive" @close-modal="toggleModal"/>
-
-  
+  <!-- <main-logo/> -->
+  <home-view/>
 
 </template>
 
 <script lang="ts" setup>
 
 import HomeView from "@/common/views/HomeView.vue";
-import { ref, watchEffect } from 'vue';
-import SignInUpModal from "./Modules/locations/components/SignInUpModal.vue";
-
-const modalActive = ref(false);
-
-const toggleModal = () => {
-  modalActive.value = !modalActive.value;
-}
-
-watchEffect(() => {
-  if (modalActive.value) {
-    document.body.classList.add('overflow-hidden');
-  } else {
-    document.body.classList.remove('overflow-hidden');
-  }
-});
 
 </script>
